refactor(product): move views into dojo/product/ui/ [product Phase 5]
