+++ dojo/asset/urls.py
@@ -2,7 +2,7 @@ from django.conf import settings
 from django.urls import re_path
 
 from dojo.engagement.ui import views as dojo_engagement_views
-from dojo.product import views
+from dojo.product.ui import views
 from dojo.utils import redirect_view
 
 # TODO: remove the else: branch once v3 migration is complete
